Fix Bug In Update Salary Method Causing Undefined ID During Edit
Seems like there's a little bug upon clicking edit from the list on a previously edited salary.

The id in the path ends up as undefined.

We can fix this is by changing updateSalary() by ensuring the ID is assigned and saved correctly to localStorage.

Otherwise, the salary will be saved in localStorage without an id.